refactor: Update BuildContext components to use boolean values instead of strings
