+++ sui/core/types.go
@@ -47,7 +47,7 @@ type Page struct {
 
 // BuildContext is the struct for the build context
 type BuildContext struct {
-	components    map[string]string
+	components    map[string]bool
 	jitComponents map[string]bool
 	sequence      int
 	doc           *goquery.Document
